finished 1.0
