+++ src/app/layout.tsx
@@ -5,8 +5,7 @@ import { Inter } from 'next/font/google'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Color Matcher',
 }
 
 export default function RootLayout({
@@ -16,6 +15,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
+       <link rel="icon" href="/colour.png" />
       <body className={inter.className}>{children}</body>
     </html>
   )
