+++ src/functions/index.ts
@@ -1,0 +1,4 @@
+import { hslToHex } from "./hslToHex";
+import { roundHSLA } from "./roundHSLA";
+
+export { hslToHex, roundHSLA };
